requirements.txt
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==1.11.29
django-crispy-forms==1.9.0
django-formtools==2.1
django-haystack==2.8.1
django-import-export==2.0.2
django-reversion==3.0.7
DjangoUeditor==1.8.143
et-xmlfile==1.0.1
future==0.15.2
httplib2==0.9.2
idna==2.9
jdcal==1.4.1
jieba==0.42.1
Markdown==3.2.1
MarkupPy==1.14
mysqlclient==1.4.6
odfpy==1.4.1
openpyxl==3.0.3
Pillow==7.1.1
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
setuptools-scm==3.5.0
simplejson==3.17.0
six==1.10.0
tablib==1.1.0
urllib3==1.25.9
Whoosh==2.7.4
xlrd==1.2.0
xlwt==1.3.0
